+++ CHANGELOG.md
@@ -158,6 +158,7 @@ Modifications can only be done via builders.
 - Suppliers must not be available from the beginning
 - Rework initialization of suppliers
 - Rename replaceNode into addBluePrintOverlay. Add Scope.addOrReplaceNode.
+- Throw if a node is replaced with one with invalid suppliers
 
 ### Fixed